+++ build.gradle.kts
@@ -1,6 +1,7 @@
 plugins {
     kotlin("jvm") version "1.9.0"
-    `maven-publish`
+
+    `java-library`
 }
 
 group = "nl.mmathijs"
@@ -15,33 +16,17 @@ repositories {
 }
 
 dependencies {
+    api(kotlin("stdlib-jdk8"))
+
     testImplementation(kotlin("test"))
+
     implementation("com.acmerobotics.roadrunner:core:0.5.6")
 }
 
+java {
+    withSourcesJar()
+}
 
 tasks.test {
     useJUnitPlatform()
-}
-
-kotlin {
-    jvmToolchain(8)
-}
-
-publishing {
-    repositories {
-        maven {
-            name = "GitHubPackages"
-            url = uri("https://maven.pkg.github.com/gamermathijs/rabid-rider")
-            credentials {
-                username = System.getenv("GITHUB_ACTOR")
-                password = System.getenv("GITHUB_TOKEN")
-            }
-        }
-    }
-    publications {
-        create<MavenPublication>("mavenJava") {
-            from(components["java"])
-        }
-    }
 }
